requirements.txt
python-telegram-bot
google-generativeai
google-genai
python-dotenv